Additional changes to COD cli from original
**Disabled github actions workflow: test-bot.yml** On the homebrew-gds repo
(https://github.com/alphagov/homebrew-gds) which this repo was forked from, it
appears that the test-bot.yml github action has been broken for at least a
year. Since we are unsure if this action ever worked, or if it was a work in
progress that has now been abandoned - we have commented it out.  The `brew
test-bot command` is usually used to test the the formula using native brew
commands and linters. However much of the linting is currently broken, and the
brew test-bot is unable to clone the cod-cli repo, as this is a private repo
that can only be reached by SSH. This problem is explained in more detail
below.

**SSH vs Https** Installing the cod cli via brew ( you can inspect the formula
here https://github.com/GovWifi/homebrew-cod/blob/master/Formula/cod-cli.rb)
requires access to the cod-cli repo over SSH. This is also how the gds-cli
formula works. This is fine if a human user is pulling the code down as part of
a brew install on their local machine, as they will have an SSH key associated
to their account. However if a github action is pulling down the code, then it
cannot use an SSH key, it MUST use a github token.

We experimented with setting a github token as value for
HOMEBREW_GITHUB_API_TOKEN environment variable as suggested by the brew
test-bot help page https://docs.brew.sh/Manpage#test-bot-options-formula
(please see line 33 of
https://github.com/alphagov/homebrew-gds/blob/master/Formula/gds-cli.rb).
However we later discovered that Github tokens only work over https NOT SSH. So
it does not seem possible to use the `brew test-bot` command to validate our
current formula.

**Final decision and possible future feature** Presently the cod-cli compiles
and works fine without this extra testing step, and we are unable to dedicate
any more developer time to fixing the broken action. However, at some future
point we may revisit this work, and perhaps refactor the formula or rewrite it
in another language which offers more flexibility than ruby around cloning
repos using https.